+++ RandomSkunk.Results.SourceGenerators/ResultTupleExtensionsGeneratorExtensions.cs
@@ -756,8 +756,7 @@ public static class ResultTupleExtensions
 
     public static StringBuilder AppendGetNonSuccessErrorMethod(this StringBuilder code)
     {
-        return code.Append(@"    [StackTraceHidden]
-    private static Error GetNonSuccessError(params IResult[] results)
+        return code.Append(@"    private static Error GetNonSuccessError(params IResult[] results)
     {
         var errors = results
             .Where(r => !r.IsSuccess)