+++ RandomSkunk.Results/Maybe{T}.cs
@@ -22,7 +22,7 @@ public partial struct Maybe<T> : IResult<T>, IEquatable<Maybe<T>>
         _error = null;
     }
 
-    private Maybe(bool none, Error? error = null)
+    private Maybe(bool none, Error? error, bool? setStackTrace)
     {
         if (none)
         {
@@ -34,7 +34,12 @@ public partial struct Maybe<T> : IResult<T>, IEquatable<Maybe<T>>
         {
             _outcome = _failOutcome;
             _value = default;
-            _error = FailResult.InvokeReplaceErrorIfSet(error ?? new Error(setStackTrace: true));
+
+            _error = error ?? new Error();
+            if (_error.StackTrace is null && (setStackTrace ?? FailResult.SetStackTrace))
+                _error = _error with { StackTrace = FilteredStackTrace.Create() };
+
+            _error = FailResult.InvokeReplaceErrorIfSet(_error);
             FailResult.InvokeCallbackIfSet(_error);
         }
     }
@@ -111,15 +116,19 @@ public partial struct Maybe<T> : IResult<T>, IEquatable<Maybe<T>>
     /// Creates a <c>None</c> result.
     /// </summary>
     /// <returns>A <c>None</c> result.</returns>
-    public static Maybe<T> None() => new(none: true);
+    public static Maybe<T> None() => new(none: true, null, null);
 
     /// <summary>
     /// Creates a <c>Fail</c> result with the specified error.
     /// </summary>
     /// <param name="error">An error that describes the failure. If <see langword="null"/>, a default error is used.</param>
+    /// <param name="setStackTrace">Whether to set the stack trace of the error to the current location. If
+    ///     <see langword="null"/> or not provided, the value of the <see cref="FailResult.SetStackTrace"/> property is used
+    ///     instead.</param>
     /// <returns>A <c>Fail</c> result.</returns>
     [StackTraceHidden]
-    public static Maybe<T> Fail(Error? error = null) => new(none: false, error);
+    public static Maybe<T> Fail(Error? error = null, bool? setStackTrace = null) =>
+        new(none: false, error, setStackTrace);
 
     /// <summary>
     /// Creates a <c>Fail</c> result.
@@ -130,6 +139,9 @@ public partial struct Maybe<T> : IResult<T>, IEquatable<Maybe<T>>
     /// <param name="errorIdentifier">The optional identifier of the error.</param>
     /// <param name="errorTitle">The optional title for the error. If <see langword="null"/>, then "Error" is used instead.
     ///     </param>
+    /// <param name="setStackTrace">Whether to set the stack trace of the error to the current location. If
+    ///     <see langword="null"/> or not provided, the value of the <see cref="FailResult.SetStackTrace"/> property is used
+    ///     instead.</param>
     /// <returns>A <c>Fail</c> result.</returns>
     [StackTraceHidden]
     public static Maybe<T> Fail(
@@ -137,8 +149,9 @@ public partial struct Maybe<T> : IResult<T>, IEquatable<Maybe<T>>
         string errorMessage = Error.DefaultFromExceptionMessage,
         int? errorCode = ErrorCodes.CaughtException,
         string? errorIdentifier = null,
-        string? errorTitle = null) =>
-        Fail(Error.FromException(exception, errorMessage, errorCode, errorIdentifier, errorTitle));
+        string? errorTitle = null,
+        bool? setStackTrace = null) =>
+        Fail(Error.FromException(exception, errorMessage, errorCode, errorIdentifier, errorTitle), setStackTrace);
 
     /// <summary>
     /// Creates a <c>Fail</c> result.
@@ -149,8 +162,9 @@ public partial struct Maybe<T> : IResult<T>, IEquatable<Maybe<T>>
     /// <param name="errorTitle">The optional title for the error. If <see langword="null"/>, then "Error" is used instead.
     ///     </param>
     /// <param name="innerError">The optional error that is the cause of the current error.</param>
-    /// <param name="stackTrace">The optional stack trace. If <see langword="null"/>, then a generated stack trace is used.
-    ///     </param>
+    /// <param name="setStackTrace">Whether to set the stack trace of the error to the current location. If
+    ///     <see langword="null"/> or not provided, the value of the <see cref="FailResult.SetStackTrace"/> property is used
+    ///     instead.</param>
     /// <returns>A <c>Fail</c> result.</returns>
     [StackTraceHidden]
     public static Maybe<T> Fail(
@@ -159,22 +173,22 @@ public partial struct Maybe<T> : IResult<T>, IEquatable<Maybe<T>>
         string? errorIdentifier = null,
         string? errorTitle = null,
         Error? innerError = null,
-        string? stackTrace = null) =>
-        Fail(new Error(errorMessage, errorTitle, setStackTrace: stackTrace is null)
-        {
-            StackTrace = stackTrace,
-            ErrorCode = errorCode,
-            Identifier = errorIdentifier,
-            InnerError = innerError,
-        });
+        bool? setStackTrace = null) =>
+        Fail(
+            new Error(errorMessage, errorTitle)
+            {
+                ErrorCode = errorCode,
+                Identifier = errorIdentifier,
+                InnerError = innerError,
+            },
+            setStackTrace);
 
     /// <summary>
     /// Creates a maybe from the specified value.
     /// </summary>
     /// <param name="value">The value. Can be <see langword="null"/>.</param>
     /// <returns>A <c>Success</c> result if <paramref name="value"/> is not null; otherwise, a <c>None</c> result.</returns>
-    public static Maybe<T> FromValue(
-        T? value) =>
+    public static Maybe<T> FromValue(T? value) =>
         value is not null
             ? Success(value)
             : None();