+++ RandomSkunk.Results/Result.cs
@@ -12,7 +12,7 @@ public partial struct Result : IResult<DBNull>, IEquatable<Result>
     private readonly int _outcome;
     private readonly Error? _error;
 
-    private Result(bool success, Error? error = null)
+    private Result(bool success, Error? error, bool? setStackTrace)
     {
         if (success)
         {
@@ -22,7 +22,12 @@ public partial struct Result : IResult<DBNull>, IEquatable<Result>
         else
         {
             _outcome = _failOutcome;
-            _error = FailResult.InvokeReplaceErrorIfSet(error ?? new Error(setStackTrace: true));
+
+            _error = error ?? new Error();
+            if (_error.StackTrace is null && (setStackTrace ?? FailResult.SetStackTrace))
+                _error = _error with { StackTrace = FilteredStackTrace.Create() };
+
+            _error = FailResult.InvokeReplaceErrorIfSet(_error);
             FailResult.InvokeCallbackIfSet(_error);
         }
     }
@@ -77,15 +82,18 @@ public partial struct Result : IResult<DBNull>, IEquatable<Result>
     /// Creates a <c>Success</c> result.
     /// </summary>
     /// <returns>A <c>Success</c> result.</returns>
-    public static Result Success() => new(success: true);
+    public static Result Success() => new(success: true, null, null);
 
     /// <summary>
     /// Creates a <c>Fail</c> result with the specified error.
     /// </summary>
     /// <param name="error">An error that describes the failure. If <see langword="null"/>, a default error is used.</param>
+    /// <param name="setStackTrace">Whether to set the stack trace of the error to the current location. If
+    ///     <see langword="null"/> or not provided, the value of the <see cref="FailResult.SetStackTrace"/> property is used
+    ///     instead.</param>
     /// <returns>A <c>Fail</c> result.</returns>
     [StackTraceHidden]
-    public static Result Fail(Error? error = null) => new(success: false, error);
+    public static Result Fail(Error? error = null, bool? setStackTrace = null) => new(success: false, error, setStackTrace);
 
     /// <summary>
     /// Creates a <c>Fail</c> result.
@@ -96,6 +104,9 @@ public partial struct Result : IResult<DBNull>, IEquatable<Result>
     /// <param name="errorIdentifier">The optional identifier of the error.</param>
     /// <param name="errorTitle">The optional title for the error. If <see langword="null"/>, then "Error" is used instead.
     ///     </param>
+    /// <param name="setStackTrace">Whether to set the stack trace of the error to the current location. If
+    ///     <see langword="null"/> or not provided, the value of the <see cref="FailResult.SetStackTrace"/> property is used
+    ///     instead.</param>
     /// <returns>A <c>Fail</c> result.</returns>
     [StackTraceHidden]
     public static Result Fail(
@@ -103,8 +114,9 @@ public partial struct Result : IResult<DBNull>, IEquatable<Result>
         string errorMessage = Error.DefaultFromExceptionMessage,
         int? errorCode = ErrorCodes.CaughtException,
         string? errorIdentifier = null,
-        string? errorTitle = null) =>
-        Fail(Error.FromException(exception, errorMessage, errorCode, errorIdentifier, errorTitle));
+        string? errorTitle = null,
+        bool? setStackTrace = null) =>
+        Fail(Error.FromException(exception, errorMessage, errorCode, errorIdentifier, errorTitle), setStackTrace);
 
     /// <summary>
     /// Creates a <c>Fail</c> result.
@@ -115,8 +127,9 @@ public partial struct Result : IResult<DBNull>, IEquatable<Result>
     /// <param name="errorTitle">The optional title for the error. If <see langword="null"/>, then "Error" is used instead.
     ///     </param>
     /// <param name="innerError">The optional error that is the cause of the current error.</param>
-    /// <param name="stackTrace">The optional stack trace. If <see langword="null"/>, then a generated stack trace is used.
-    ///     </param>
+    /// <param name="setStackTrace">Whether to set the stack trace of the error to the current location. If
+    ///     <see langword="null"/> or not provided, the value of the <see cref="FailResult.SetStackTrace"/> property is used
+    ///     instead.</param>
     /// <returns>A <c>Fail</c> result.</returns>
     [StackTraceHidden]
     public static Result Fail(
@@ -125,14 +138,15 @@ public partial struct Result : IResult<DBNull>, IEquatable<Result>
         string? errorIdentifier = null,
         string? errorTitle = null,
         Error? innerError = null,
-        string? stackTrace = null) =>
-        Fail(new Error(errorMessage, errorTitle, setStackTrace: stackTrace is null)
-        {
-            StackTrace = stackTrace,
-            ErrorCode = errorCode,
-            Identifier = errorIdentifier,
-            InnerError = innerError,
-        });
+        bool? setStackTrace = null) =>
+        Fail(
+            new Error(errorMessage, errorTitle)
+            {
+                ErrorCode = errorCode,
+                Identifier = errorIdentifier,
+                InnerError = innerError,
+            },
+            setStackTrace);
 
     /// <inheritdoc/>
     public bool Equals(Result other) =>